requirements.txt
ultralytics
torch
opencv-python
numpy
pyyaml
google-cloud-storage
tqdm

